pyproject.toml
[project]
name = "intensifyer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Javier Ferrer", email = "javier.f.g@um.es" }]
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.0",
    "imageio>=2.37.0",
    "imageio-ffmpeg>=0.6.0",
    "loguru>=0.7.3",
    "numpy>=2.2.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "python-telegram-bot>=21.10",
]

[project.scripts]
intensifyer = "intensifyer:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
